+++ app/widgets.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Part of the Platform application.
  *
@@ -10,10 +11,10 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Platform
- * @version    5.0.0
+ * @version    6.0.0
  * @author     Cartalyst LLC
  * @license    Cartalyst PSL
- * @copyright  (c) 2011-2016, Cartalyst LLC
+ * @copyright  (c) 2011-2017, Cartalyst LLC
  * @link       http://cartalyst.com
  */
 
@@ -27,4 +28,4 @@
 |
 */
 
-Widget::map('hello', 'App\Widgets\HelloWidget@show');
+# Widget::map('hello', 'App\Widgets\HelloWidget@show');