+++ config/platform/config.php
@@ -10,7 +10,7 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Platform
- * @version    5.0.0
+ * @version    6.0.0
  * @author     Cartalyst LLC
  * @license    Cartalyst PSL
  * @copyright  (c) 2011-2016, Cartalyst LLC
@@ -75,7 +75,7 @@ return [
         |
         */
 
-        'copyright' => '&copy; 2011-2016, Cartalyst LLC',
+        'copyright' => '&copy; 2011-2017, Cartalyst LLC',
 
         /*
         |--------------------------------------------------------------------------