+++ CHANGELOG.md
@@ -11,6 +11,18 @@ We do not give estimated times for completion on `Accepted` Proposals.
 
 ---
 
+### v6.0.0 - 2017-02-24
+
+- Upgraded to Laravel 5.3.
+- Upgraded to Themes v4 package.
+- Upgraded to Extensions v3 package.
+- Upgraded to Data Grid v4.
+- Upgraded to Sentinel Social v3.
+
+### v5.0.0 - 2016-08-02
+
+- New media management features.
+
 ### v4.0.0 - 2015-07-13
 
 - Upgraded to Laravel 5.1.